Week 3 assignment
Create your navigation as an unordered list using <nav>, <ul> and <li>.

Style all four pseudo-classes for the links in your stylesheet. Use
selector order or a class or ID to apply the link styles only to links
in your <nav>

Create the remainder of your pages with the nav in them and link them
to one another. (You don't need the full page content in each page,
just the navigation that links to all other pages.)

Make sure that all your links are working to all other pages, from all
your pages.